fix: bad permissions for Nerdfont directories
Some days ago, I stablished some permissions in a task to supress a
Warning on the Ansible linter. As I did not pay much attention, the
permissions where wrong, removing my apps access to those fonts, with
the consequent visualization errors.
I changed the permissions to the correct ones, fixing the problem.
In addition, the way of installing Docker in the machines has changed
from the default Ubuntu packages, to the geerlingguy.docker ansible
role.